requirements.txt
pandas
sentence-transformers
scikit-learn
joblib
python-dotenv
groq
fastapi
uvicorn[standard]
pydantic
